+++ src/appleseed/renderer/kernel/lighting/sppm/sppmparameters.cpp
@@ -139,8 +139,8 @@ void SPPMParameters::print() const
         "  rr min path len. %s",
         pretty_uint(m_light_photon_count).c_str(),
         pretty_uint(m_env_photon_count).c_str(),
-        m_photon_tracing_max_path_length == ~0 ? "infinite" : pretty_uint(m_photon_tracing_max_path_length).c_str(),
-        m_photon_tracing_rr_min_path_length == ~0 ? "infinite" : pretty_uint(m_photon_tracing_rr_min_path_length).c_str());
+        m_photon_tracing_max_path_length == size_t(~0) ? "infinite" : pretty_uint(m_photon_tracing_max_path_length).c_str(),
+        m_photon_tracing_rr_min_path_length == size_t(~0) ? "infinite" : pretty_uint(m_photon_tracing_rr_min_path_length).c_str());
 
     RENDERER_LOG_INFO(
         "sppm path tracing settings:\n"
@@ -150,8 +150,8 @@ void SPPMParameters::print() const
         "  alpha            %s\n"
         "  max photons/est. %s\n"
         "  dl light samples %s",
-        m_path_tracing_max_path_length == ~0 ? "infinite" : pretty_uint(m_path_tracing_max_path_length).c_str(),
-        m_path_tracing_rr_min_path_length == ~0 ? "infinite" : pretty_uint(m_path_tracing_rr_min_path_length).c_str(),
+        m_path_tracing_max_path_length == size_t(~0) ? "infinite" : pretty_uint(m_path_tracing_max_path_length).c_str(),
+        m_path_tracing_rr_min_path_length == size_t(~0) ? "infinite" : pretty_uint(m_path_tracing_rr_min_path_length).c_str(),
         pretty_scalar(m_initial_radius_percents, 3).c_str(),
         pretty_scalar(m_alpha, 1).c_str(),
         pretty_uint(m_max_photons_per_estimate).c_str(),
